requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj-database-url
dj-static
Django
django-allauth
django-environ
django-mathfilters
django-tz-detect
django-widget-tweaks
easy-colors
gunicorn
humanize
idna
oauthlib
packaging
pillow
psycopg2-binary
pycparser
PyJWT
python-decouple
python3-openid
pytz
requests
requests-oauthlib
send-mail
sqlparse
static3
typing_extensions
tzdata
urllib3
whitenoise